feat : 회원 관련 전체 기능 업데이트 (소셜 로그인, 로그인, 이메일 센더 미구현)

+++ server/src/main/java/com/codemouse/salog/members/entity/Member.java
@@ -15,7 +15,7 @@ import java.util.*;
 public class Member extends Auditable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int memberId;
+    private long memberId;
 
     @Column(nullable = false, unique = true)
     private String email;
